requirements.txt
amqp==1.4.6
anyjson==0.3.3
arrow==0.5.4
billiard==3.3.0.19
breadability==0.1.20
celery==3.1.17
chardet==2.3.0
contextlib2==0.4.0
cssselect==0.9.1
dj-database-url==0.3.0
Django==1.7.1
django-celery==3.1.16
django-cors-headers==1.0.0
django-postgrespool==0.3.0
django-simpleapi==1.3.1
django-taggit==0.12.2
docopt==0.6.2
gnureadline==6.3.3
gunicorn==19.1.1
iso8601==0.1.10
kombu==3.0.24
lxml==3.4.1
mock==1.0.1
psycopg2==2.5.4
python-dateutil==2.4.0
pytz==2014.10
PyYAML==3.11
requests==2.5.1
responses==0.3.0
schematics==1.0.post0
six==1.9.0
SQLAlchemy==0.9.8
vcrpy==1.1.4
w3lib==1.10.0
whitenoise==1.0.6
wrapt==1.10.2
